+++ data-raw/list_supported_datasets_ff.R
@@ -1,7 +1,8 @@
-# This script generates the hardcoded tibble used in list_supported_types_ff().
-# Run this script if the list of available Fama-French datasets needs to be
-# updated. The output of datapasta::tribble_paste() should replace the
-# tribble() call in R/list_supported_types.R.
+# This script generates the hardcoded tibble used in
+# list_supported_datasets_ff(). Run this script if the list of available
+# Fama-French datasets needs to be updated. The output of
+# datapasta::tribble_paste() should replace the tribble() call in
+# R/list_supported_datasets.R.
 #
 # Requires: frenchdata, dplyr, datapasta